ecommerce/services/auth_service/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pillow==10.4.0
python-dotenv==1.0.1
gunicorn==22.0.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
Faker==30.3.0

# ── Linting ──────────────────────────────────────────────────────────────────
ruff==0.7.1
